+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/Ktor/base/AbstractRequestCallFactory.kt
@@ -4,9 +4,8 @@ import com.github.insanusmokrassar.TelegramBotAPI.bot.Ktor.KtorCallFactory
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
 import io.ktor.client.HttpClient
 import io.ktor.client.call.HttpClientCall
-import io.ktor.client.call.call
-import io.ktor.client.request.accept
-import io.ktor.client.request.url
+import io.ktor.client.request.*
+import io.ktor.client.statement.HttpStatement
 import io.ktor.http.ContentType
 import io.ktor.http.HttpMethod
 import kotlin.collections.set
@@ -17,20 +16,23 @@ abstract class AbstractRequestCallFactory : KtorCallFactory {
         client: HttpClient,
         baseUrl: String,
         request: Request<T>
-    ): HttpClientCall? {
+    ): HttpStatement? {
         val preparedBody = prepareCallBody(client, baseUrl, request) ?: return null
 
-        return client.call {
-            url(
-                methodsCache[request.method()] ?: "$baseUrl/${request.method()}".also {
-                    methodsCache[request.method()] = it
-                }
-            )
-            method = HttpMethod.Post
-            accept(ContentType.Application.Json)
+        return HttpStatement(
+            HttpRequestBuilder().apply {
+                url(
+                    methodsCache[request.method()] ?: "$baseUrl/${request.method()}".also {
+                        methodsCache[request.method()] = it
+                    }
+                )
+                method = HttpMethod.Post
+                accept(ContentType.Application.Json)
 
-            body = preparedBody
-        }
+                body = preparedBody
+            },
+            client
+        )
     }
 
     protected abstract fun <T : Any> prepareCallBody(