+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/RequestsExecutor.kt
@@ -1,7 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.bot
 
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
-import kotlinx.io.core.Closeable
+import io.ktor.utils.io.core.Closeable
 
 interface RequestsExecutor : Closeable {
     /**