+++ src/jsMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/StorageFile.kt
@@ -1,7 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.utils
 
 import com.benasher44.uuid.uuid4
-import kotlinx.io.core.Input
+import io.ktor.utils.io.core.Input
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient