+++ src/jvmMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/StorageFile.kt
@@ -1,8 +1,8 @@
 package com.github.insanusmokrassar.TelegramBotAPI.utils
 
 import com.benasher44.uuid.uuid4
-import kotlinx.io.core.Input
-import kotlinx.io.streams.asInput
+import io.ktor.utils.io.core.Input
+import io.ktor.utils.io.streams.asInput
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 import java.io.File