+++ CHANGELOG.md
@@ -14,6 +14,8 @@
     * `UserForwardedMessage` -> `UserForwardInfo`
     * `ForwardedFromChannelMessage` -> `ForwardFromChannelInfo`
     * `PossiblyForwardedMessage#forwarded` field now renamed to `forwardInfo`
+* All serializers in library now are `internal`. **If you have used some of them or I have marked as internal by a
+mistake - don't hesitate to say this.**
 * Most part of requests have changed return type. They are listed below:
     <details>